Fix transport authentication errors on redirect
Add a name.Registry as the target registry for a bearerTransport. The
bearerTransport will only attach bearer tokens to requests matching the
registry's domain.

Similarly, add a target host for basicTransport authentication.

This fixes an issue where the registry would redirect to object storage,
which would reject the request due to an invalid Authorization header.