Create a separate v1.Layer abstraction (#62)
* Create a separate v1.Layer abstraction.

The number of methods on v1.Image that dealt in either "Diff ID" or "Digest" was a little silly, and carry-over from the Python library.  This change defines a v1.Layer abstraction to hold the bulk of these methods, and dramatically simplify the v1.Image interface.

Fixes: https://github.com/google/go-containerregistry/issues/59

* Incorporate @jonjohnsonjr feedback.

Largely updating comments.

A TODO added to memoize in `uncompressedLayerExtender`.

Changed the manifest computation in `uncompressedImageExtender` to better utilize the `Layer` methods.