v1.SHA256, v1.ParseConfigFile, v1.ParseManifest now take an io.Reader
Prior these functions would accept an io.ReadCloser and
always close it. This behaviour is a bit rigid if you
didn't want it to close the io.Closer - ie. reusing
an os.File with different decorators

Since these methods are no longer responsible for closing -
this hopefully leads to the lifecyle (create/close) of a
io.ReadCloser to be contained within a smaller scope.

Signed-off-by: Brenda Chan <brchan@pivotal.io>