Provide a new pkg & cmd that allows mutating images + extras (#63)
* v1.Manifest and v1.Config now have DeepCopy methods

The tooling is vendored to generate these deep copy methods

Signed-off-by: Brenda Chan <brchan@pivotal.io>

* v1.SHA256, v1.ParseConfigFile, v1.ParseManifest now take an io.Reader

Prior these functions would accept an io.ReadCloser and
always close it. This behaviour is a bit rigid if you
didn't want it to close the io.Closer - ie. reusing
an os.File with different decorators

Since these methods are no longer responsible for closing -
this hopefully leads to the lifecyle (create/close) of a
io.ReadCloser to be contained within a smaller scope.

Signed-off-by: Brenda Chan <brchan@pivotal.io>

* func GzipReadCloser() is more memory efficient

Signed-off-by: Brenda Chan <brchan@pivotal.io>

* tarball.Write will now serialize the manifest layers in the correct order

tarball.Write serialized manifest layers depending on what is returned
by v1.Image.Layers. The interface's contract says it will return the
layers sorted with the most recent one first. This is the opposite
of the OCI image spec.

This caused a few problems:
- docker load would fail to import the output tars
- creating a v1.Image from a correct tarball would have it's
  layers reversed

Signed-off-by: Brenda Chan <brchan@pivotal.io>

* tarball.LayerFromFile will return a v1.Layer backed by a tarball

Additionally supports gzip'd tarballs

Signed-off-by: Brenda Chan <brchan@pivotal.io>

* Appending layers is now possible using the mutate package

Signed-off-by: Brenda Chan <brchan@pivotal.io>

* DEPRECATED: go_repositories has been deprecated. - lets remove it

* Add a sub-command that appends tarballs to an image in a Docker registry.

Signed-off-by: Brenda Chan <brchan@pivotal.io>