Implement ko/publish
This adds a package for handling how `v1.Image`s are published for `ko`.  The `publish.Interface` contains one method, which takes a `v1.Image` and an importpath and is responsible for publishing it to some base `name.Repository` and returning a `name.Digest` suitable for substitution into the K8s yaml.

There are currently two implementations provided:
1. `NewFixed`: this is an implementation intended to aid in testing packages that need a `publish.Interface`.
1. `NewDefault`: this is an implementation that publishes to: `{chroot}/{importpath}:latest` via `remote.Write` and the default keychain.

In the future, we might consider alternate implementations of `publish.Interface` that support flatter naming schemes (e.g. to support DockerHub, with which the `NewDefault` scheme won't work).

Progress towards: #80