Add a local mode to ko.
This change adds a --local (or -L) flag to `ko`, which directs it to "publish" image references differently.  Instead of publishing the image to a registry, and substituting its digest, this change adds the option to publish the image to the docker daemon.

The scenario this is aimed at supporting is local development with minikube without a registry roundtrip.  With minikube as `kubectl config current-context`, this can load the images directly into the minikube docker daemon and then direct minikube to run the resulting templates.

Fixes: https://github.com/google/go-containerregistry/issues/133