Rename our chosen entrypoint to /ko-app
Previously we used `/app`, which is unfortunately a fairly generic name, in a location that doesn't disambiguate well.  Under most circumstance this is fine, however, I ran into a problem with a base image that was using `/app` for other purposes and this unintentionally shadowed that directory with our app.