Accept KO_DOCKER_REPO=ko.local in addition to -L
This enables users to configure the use of the local docker daemon via the `KO_DOCKER_REPO` environment variable in addition to supplying `-L` on the command-line.

There have been reports that `-L` makes it harder for teams that have a mix of minikube and K8s cluster usage to script with `ko`.  This is expected to make that somewhat simpler.

Fixes: https://github.com/google/go-containerregistry/issues/179