Avoid recomputing the DiffID of layers where we know it.
Prior to this change, the `partial` package enabled folks to implement the bare minimum interface necessary to compute the remainder of the `v1.Image` or `v1.Layer` interface.  With this change, we start to adopt a pattern that allows implementors to go beyond this bare minimum and implement additional methods and have the `partial` wrappers respect and delegate to these additional methods.

The first use case of this pattern is in `remote.Layer` where we effectively know all of the layers' `DiffID`s, but because of the "bare minimum" constraint we had previously been unable to leverage this knowledge to elide recomputing the DiffID by downloading and uncompressing the layer.

Fixes: https://github.com/google/go-containerregistry/issues/166