Change the way mounting works in go-containerregistry. (#186)
This completely changes the model for cross-repo mounting in the go-containerregistry library.  Rather than having clients of `remote.Write` pass in a set of potential mount sources, `remote.Write` is sensitive to a new `v1.Layer` wrapper type: `remote.MountableLayer`, which implements `v1.Layer` by embedding the layer it wraps and augmenting that with the name of the repository from which we should attempt to mount it, if written.  With this model, `remote.Image` now wraps all of the `v1.Layer`s it returns with the name of the repository from which it is pulling.

Fixes: https://github.com/google/go-containerregistry/issues/165