Expose a flag to flatten import paths in image names.
Add a functional option to `pkg/ko/publish` to support flat image names, and utilize this when `--flat` is passed to `ko`.  This will change an import path to something like: `{package}-{hash of full import path}`.

I have verified with this flag that I can publish to DockerHub via `ko`.

Fixes: https://github.com/google/go-containerregistry/issues/160