Fix anonymous credential fallback when using keychain credentials.
The credentials helpers used to access osxkeychain and thelike print to stdout which is promne to contain spacing characters. Trimming before comparing to the "magicNotFoundString" fixes a bug where a fallback to anonymous credentials did not work if the user was not logged into the respective registry at all.

Fixes #233