Simplify flags a bit
Removed "tags" and "untagged" filters.
Added "untag" as a behavior.

Having a set of tags as a filter seems like it would be surprising. Do
we delete all the tags for an image that matches? Just the tags in the
flag?

The default behavior remains the same (only print untagged images), but
passing --untag or -U will also print matching images and their tags.

This is kind of confusing still, though. You might expect --untag to
actually untag things, but it will only do that if you pass -U AND -D.