Add googley auth mechanisms
Going through the credential helper for everything has some drawbacks.
It's slow because we have to shell out twice every time we authenticate.
We also can't cache the tokens that get returned to us because docker's
credential helper protocol doesn't include expiration.

This reimplements a lot of the GCR credential helper logic using the
authn package's abstractions, exposed as:

google.NewEnvAuthenticator, which looks in the environment for
credentials (e.g. on GCE) and caches them.

google.NewGcloudAuthenticator, which shells out to gcloud for
credentials and caches them.

google.Keychain, which returns the env-based authenticator, falling back
to the gcloud-based authenticator if that fails.