Upload config concurrently if we can (#323)
* Upload config concurrently if we can

https://github.com/google/go-containerregistry/pull/301 introduced a
pretty big regression in upload speed. To get around this, attempt to
read the config blob before waiting for layer uploads to finish. If we
can read it, add it to the errgroup, otherwise wait for the layers to
finish before attempting to upload the config again.

In my testing, this change shaves ~3s off of ko publishing.

* Check for ErrNotComputed error