Switch ko resolution to use "futures".
Today the resolution logic in `ko` resolves all files before it starts to feed anything to `kubectl`.

This changes that logic to use channels to stream files back from the filesystem discovery process and futures to block on the resolution *in order* so that they are fed to kubectl in the order they are received.  This last bit is subtle and important because the order of the initial walk is generally expected to deterministically order how things are fed to kubectl.

This change is also a prelude to another that will make the file enumeration a continuous stream that's hooked up to a filesystem watcher like fsnotify.