ko apply --watch
This adds experimental support for fsnotify integration for `ko apply`.

With this change you can run: `ko apply --watch -f config/` and it will perform the typical deployment step, and then begin to monitor the referenced go paths from each of the yaml files for changes using github.com/mattmoor/dep-notify (a Go dependency graph library built around fsnotify).

When a change is found, only the yaml files that reference the image in question are re-applied.  In watch mode, compilation failures are ignored (so they may be remedied and picked up by the watch).