Add a caching builder.
This caching builder creates a unique build "future" per import path, and returns the result of that build future each time that import path is requested (until it is invalidated).

When we see filesystem changes that affect a particular build target, we `Invalidate(import path)` and fresh builds will be scheduled for that import path.

This also introduces a TODO to avoid republishing the same image, which is needed to fully eliminate redundant work.