Add a caching publisher.
This caching publisher creates a unique publish "future" per reference (associated with the image being published), and returns the result of that publish future each time that reference is requested (so long as it has the same image as the one we associated with the cached reference).

The result of publishing a particular v1.Image should never change, so invalidation is handled by the v1.Image supplied to `Publish()` changing (which it will given the caching build semantics).