Update v1.ImageIndex to be more usable
1. Rename RawIndexManifest to RawManifest
2. Add Image(Hash)
3. Add ImageIndex(Hash)

(1) will allow us to reuse a bunch of code that just deals with the
content-addressable aspects of manifests.

(2) and (3) will allow us to use ImageIndex in the same way as we
currently use Image, since it will know how to fetch the things that it
references. This is similar to how Image.Layers works.