Refactor remote.{Image,ImageIndex} onto Descriptor
This consolidates the remoteImage and remoteIndex logic such that it
goes through a remote.Descriptor, which does appropriate type checking
and error handling in one place.

Doing so required a bit of shuffling around, which made the ImageIndex
-> Image resolution a lot cleaner, and enabled code reuse for the
v1.ImageIndex.{Image,ImageIndex} methods for accessing child artifacts.

Skip digest checks if we get back a signed schema 1 manifest, since we'd
have to remove the signatures to get the digest to match.