Add remote.Descriptor (#408)
* Add remote.Descriptor

This resolves a mismatch between how registries actually work and our
Image/ImageIndex interfaces. Since we don't know what type of artifact a
registry stores, we need some kind of union type to represent either
of them. This allows clients to reason about what type of artifact they
got back from the registry, and efficiently "cast" it to the right
implementation. This removes the need to do fallback.

remote.Descriptor embeds a v1.Descriptor, so it contains MediaType,
Size, and Digest, but it also has a Manifest field so we can access
the raw response from the registry. This is useful for crane digest and
crane manifest, where we dont' want to interpret the response, just show
what's there.

The remote.Image and remote.Index functions have been rewritten to use
remote.Descriptor under the hood. One observable side-effect of this is
that these functions will fail earlier than before. Whereas we used to
just perform the token handshake in e.g. `remote.Image(ref)`, now we
fetch the manifest (potentially recursively). In my opinion, this is a
good thing. It's better to fail early than later. One thing this fixes
is that e.g. crane push will fail with a message about pulling instead
of pushing, since we'll fail before we try to call remote.Write. This is
better for debugging.

Doing this eager manifest fetching also allows us to return a better
error message in the case of schema 1 images.

There are downsides to this, e.g. passing around collection of v1.Image
(before calling (Image.RawManifest)) will take up more memory (the size
of the manifest, just a few KB). That really only affects
tarball.MultiWriteToFile, but the difference should be negligble.

* Move remote.fetcher

remote/image.go -> remote/descriptor.go

* Refactor remote.{Image,ImageIndex} onto Descriptor

This consolidates the remoteImage and remoteIndex logic such that it
goes through a remote.Descriptor, which does appropriate type checking
and error handling in one place.

Doing so required a bit of shuffling around, which made the ImageIndex
-> Image resolution a lot cleaner, and enabled code reuse for the
v1.ImageIndex.{Image,ImageIndex} methods for accessing child artifacts.

Skip digest checks if we get back a signed schema 1 manifest, since we'd
have to remove the signatures to get the digest to match.

* Refactor crane {manifest,digest}

This uses remote.Descriptor, so it's now more accurate, in particular,
both commands will now do the expected thing for manifests, manifests
lists, and even schema 1 images, since we just return what we got back
from the registry without trying to parse it as a v1.Image or
v1.ImageIndex.

Potentially this could break "crane digest" if users expect it to do
image resolution first, but downstream tools should know how to resolve
manifest lists anyway.

* Refactor crane copy

Remove the fallback logic in favor of remote.Get.

* Increase test coverage

Cover new remote.Get paths and refactored index paths.

* Fix comments