Resolve ports to well-known values when comparing Host values to decide whether or not to send Bearer Authorization header.
Previously the code considered 'https://registry.example.com/' functionally different to 'https://registry.example.com:443', and similarly for 'http' and port 80.  This change always resolves Host values (which in Go may or may not be a combination of host and port) to include the host and port based on the scheme being used such that these two examples are considered functionally equivalent.