remote: Call Get() directly from remote.Image()
The descriptor.Image() code handles docker v2schema1 images cleanly, but
the Image code resricts the list, causing certain repositories to return
ManifestUnknown for images that exist.

By using the Get function, we reduce code slightly and handle schema1
images with an error that doesn't appear to imply that the image does
not exist.