Add crane.Option and make --insecure global
I realized I want to be able to inject my own transport for crane, but
that's going to be impossible if we only accept name options, so this
wraps this options (again, sigh) in a way that lets us extend to other
packages.

This also drops -i as an option for insecure (breaking, but so briefly).