Add partial.Descriptor (#654)
This consolidates a bit of logic for generating a v1.Descriptor, but
also works around some of the nastiness that the partial package
introduces.

Because partial wraps layers and images with an extra level of
implementation, it shadows any methods that the underlying object
implements, so we can't expose "shortcuts". E.g. for tarball, we may
want to expose the URLs from LayerSources to callers, but the partial
wrapping gets in the way.

This is going to be a useful pattern in general unless we get rid of the
partial package entirely. Accessor methods that can introspect the
(private) partial structs will be necessary if we want to support
"optional" methods like UncompressedSize or Descriptor.