Change signature of remote.WriteLayer (#662)
This is breaking, but I suspect there aren't many consumers. We don't
actually need a name.Digest in order to upload a layer, it was just
easier because the writer{} took a ref. I restructured commitImage to
take the ref directly instead of stashing it away in the writer{}.

If you wanted to do a streaming layer upload with WriteLayer, you'd have
to conjure up a fake name.Digest. That is no longer required, since all
we really need is the repository.