Add tests for gcrane.Copy[Repository] (#660)
* Split off TLS portion of registry.TLS

This enables spoofing of a domain for things that aren't pkg/registry,
e.g. a fake implementation of GCR.

* Add tests for gcrane.Copy[Repository]

I needed to construct a fake GCR that implements GCR's idiosyncratic
tags listing API and falls back to pkg/registry for the distribution
impelementation.

* Implement json.Marshal for google.ManifestInfo

When we Unmarshal this, we transform it from a raw form into a
well-typed form. We need to do the opposite if we want this to
roundtrip.