Add WithClient to daemon.Image
Fixes #139

Rather than wrapping the docker client options (as proposed in #139),
this just allows you to pass in an entire client implementation. This
makes testing a nicer, since we can pass in a MockImageSaver, and
callers can configure the client however they'd like without us having
to keep up with the docker client dependency.

By default, we'll still use the same options (client.FromEnv) to reach
the daemon (for compatibility).