Preserve Descriptor in remote where possible
If we resolve an index to a child, we can potentially lose a lot of
information, e.g. annotations or anything in the platform that's not in
the config, which only has os/arch.

If we return a layer from an image, we can look up the original
descriptor from that image to return as well, which might include
annotations or urls. Previously, the Mountable wrapping made this not
work.