Cache the K8s keyring across k8schains. (#705)
After adding the ECR support, we started seeing extremely slow reconciles, even off AWS.  The logic that elides this is the `Enabled()` check, which is the only thing that should run off AWS, the problem (unconfirmed) is that we run this every time we create a k8schain (which Knative was doing per-Reconciliation) because historically this has been relatively cheap.

The fix (unconfirmed) is to simply cache the underlying `credentialprovider.DockerKeyring` across invocations, which seems worth doing regardless because it takes in zero context from the caller.  We do this in a `sync.Once` (vs. `init`)to avoid this hit when the keychain is unused, but linked.