Support MultiWrite for multiple repos
This naively segments the input map by destinationn repository and
processes each repo's uploads separately, in parallel. This could be
improved to increase throughput by maintaining a map of repo to writer,
but this complicates blob uploads which had otherwise lost a reference
to the repo they were being uploaded to.

I think this form is also easier to read and comprehend, and is I think
unlikely to cause serious slowness, especially compared to the status
quo before MultiWrite.

In any case, the method signature remains the same and future
optimizations are possible without changing that.