Set default registry when building the registry object
Moves registry name defaulting to construction rather than defaulting in
registry methods. This allows for a user to override the configured
default via name options.

Note that we still rewrite docker.io to index.docker.io if it is
explictly provided. This also means that if the default is set to
docker.io by a user and no registry is provided, we will rewrite the
defeault to index.docker.io.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>