Set default tag when building tag object
Moves tag defaulting to construction of the tag object rather than
defaulting in tag methods.

Note that we assume a default tag is valid in the eyes of the user and
do not check the default for validity if it is overridden. This is
consistent with the current behavior as defaulting was previously
happening much later than the check for a valid tag.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>