Add ability to override default registry and tag in pkg/name (#824)
* Add name Option functions

Adds a name OptionFn type that allows for overriding default option
values. Also introduces two new option fields, defaultRegistry and
defaultTag. These are set to the previously used defaults unless
overridden by their respective OptionFns.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>

* Set default registry when building the registry object

Moves registry name defaulting to construction rather than defaulting in
registry methods. This allows for a user to override the configured
default via name options.

Note that we still rewrite docker.io to index.docker.io if it is
explictly provided. This also means that if the default is set to
docker.io by a user and no registry is provided, we will rewrite the
defeault to index.docker.io.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>

* Set default tag when building tag object

Moves tag defaulting to construction of the tag object rather than
defaulting in tag methods.

Note that we assume a default tag is valid in the eyes of the user and
do not check the default for validity if it is overridden. This is
consistent with the current behavior as defaulting was previously
happening much later than the check for a valid tag.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>

* Add ParseReference tests for registry and tag defaulting

Adds tests for paring full image references with registry and tag
defaulting.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>