Expose transport.NewUserAgent
This allows you to supply a customer user agent string that will be
prepended before the go-containerregistry portion.

This uses go module build info to determine the version of
go-containerregistry that you have pulled in, which will be included in
the user agent as "go-containerregistry/${VERSION}". If we cannot get
the version, this is just "go-containerregistry".