Set transport.Version via ldflags for release
For most things consuming this package, they will depend on an actual
released version of go-containerregistry. When we release
go-containerregistry itself, the binaries are built at the same version
as the library, which means that they don't depend on an actual released
version of go-containerregistry, so the version in debug info is just
"(devel)". For our own release, we will just set it to the tag.