crane.Append: Use stream.Layer for pipes
It's annoying that `crane append` is hard to use with one-liners, which
is the secret goal of most crane commands. We now stat the layer files
and use a streaming layer for anything that's not a regular file, which
includes named pipes, which fail if we try to open them multiple times.

Now you can do this:

$ crane append -f <(tar -c some-dir/ -f -) -t $IMAGE