Peek at gzip header bytes
Instead of just throwing away the io.Reader, gzip.Is now uses a
bufio.Reader to Peek at the first two bytes and returns the bufio.Reader
for the caller to reuse. This allows us to avoid resetting streams in
certain situations, for example, in tarball.Layer.

In tarball.Layer, I've added a "reopener" to replace the given Opener,
which will reuse the bufio.Reader returned by gzip.Is for the first
invocation, and just call the original opener for subsequent
invocations.