crane auth get: convert back to credentials
Before, this would just print the AuthConfig that we get back from the
config file read. This is often what we want, but it prevents crane from
acting as a credential helper itself.

By converting back to the credential helper structs, we can use crane as
a meta-credential-helper that can multiplex over anything in your config
file.