Add mutate.RebaseIndex
This method rebases all images in an ImageIndex on corresponding
platform-specific images in a new base ImageIndex.

If the new base image includes a different set of platform-specific
images than the old base image, the intersection of the two is used to
produce the rebased image index. If the intersection is the empty set,
rebasing fails.