Give the ping context a timeout. (#1163)
Debugging why requests to an insecure registry seem to hang, I noticed that there are 3+ requests to the https endpoint before it falls back to http (at that point my logs truncate because the test was over).  The time between the request being issued and completing in the roundtripper was ~30s each.

This change gives each ping a generous 5s timeout to complete each request.