update crane mutate annotation/label args to allow commas in label values
With the current argument type, a common in a label value (a reasonable thing to have)
will be split and parsed as a separate label, and the client errors. This commit updates
the flag to StringToStringVarP so we start with a map[string]string off the bat, and do not
need to do special parsing beyond checking for empty strings.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>