Make crane export more ergonomic
Default second arg to stdout so I don't have to type "-".

For images with a single layer, avoid calling mutate.Extract to flatten
the filesystem, instead we just write the uncompressed bytes to stdout.
In order to make that work, partial.Compressed now actually detects
whether the payload is gzipped or not when calling Uncompressed instead of just
assuming all blobs are compressed (and failing).