Add `--image-refs` to `crane push`.
This adds an `--image-refs` flag to `crane push` similar to the flag
I recently added to `ko`, which emits a file containing the digests
of the various images published by the command.

The intention of this flag is to work well with things like Tekton
Chains' `IMAGES` result for attesting to how a bunch of images were
produced.