Add EOF to remote default retry predicate
We see EOF when the server closes a connection without sending an
appropriate "Connection: close" header. The http Client attempts to
reuse a connection, but it fails with EOF because that connection is
closed. For idempotent requests, the http client will handle this for
us, but because this is a POST request, we just get the EOF back.