Include blob existence checks in retries (#1307)
Currently we just fail outright if the registry returns a 500 during a
blob HEAD because it's not included in the retry logic. This is not
ideal, especially in cases where an upload actually succeeded but we
never got a response from the server so we're re-uploading a blob for no
reason.