Wrapper transport no longer checks /v2 endpoint
When the calling the function NewWithContext, if the RoundTripper
provided is of the type transport.Wrapper, the function will assume that
the caller did the needed due diligence to ensure the registry supports
OCI v2. In this case the function will only need to return back the
provided RoundTripper without any further change.

Signed-off-by: João Pereira <joaod@vmware.com>